requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
echome-sdk>=0.4.0
idna==3.2
requests==2.26.0
tabulate==0.8.9
urllib3==1.26.7
